requirements.txt
# Core Dependencies
gymnasium>=0.29.0
highway-env>=1.8.2
stable-baselines3>=2.2.1
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
tensorboard>=2.14.0
tqdm>=4.66.0

# Optional: Web visualization
streamlit>=1.28.0
plotly>=5.17.0

# Development
pytest>=7.4.0
black>=23.9.0
